v2.3.16: Build system hardening and canvas test stabilization
- Added AIService.cpp and ExportService.cpp to src/CMakeLists.txt
- Removed 3 duplicate source entries: EncryptionService.cpp (duplicate),
  FileWatcher.cpp (already in PlatformServices.cpp),
  ExtensionTelemetry.cpp (already in TracingService.cpp)
- Fixed -Werror compiler warnings: unused parameters in AIService.cpp,
  unused private fields in AIService.h/ExportService.h,
  sign-conversion in ExportService.cpp
- Canvas test GLOB build with filter-exclusion in tests/CMakeLists.txt
- Phase 12 canvas tests: 44/44 pass
- Phase 13 canvas collab tests: 48/48 pass
- Main markamp binary links successfully
- Phase 50 release preparation components added

+++ src/core/ExportService.cpp
@@ -240,12 +240,12 @@ auto ExportService::generate_toc(const std::string& markdown_source) -> std::str
     {
         if (line.starts_with("#"))
         {
-            int level = 0;
-            while (level < static_cast<int>(line.size()) && line[level] == '#')
+            size_t level = 0;
+            while (level < line.size() && line[level] == '#')
                 ++level;
             if (level > 0 && level <= 6)
             {
-                auto heading = line.substr(static_cast<size_t>(level));
+                auto heading = line.substr(level);
                 // Trim leading space.
                 if (!heading.empty() && heading[0] == ' ')
                     heading = heading.substr(1);
